Added initial code for interaction with Elasticsearch (WIP)
Needs to be more robust for error handling, logging etc.
Needs more tests. Current rough tests depend on a running indexer docker
and were specific to whatever data was in my running copy, so they need
to be refined - need a docker instance for indexer added to integration
and many more tests.