Change meaning and calculation of status
Submission now has an aggregatedDepositStatus rather than a status that
attempts to calculate compliance. It simply takes all current Deposit
records and infers an overall Submission status
(aggregatedDepositStatus).

The options for Deposit and Submission status have also been reduced to
reflect that Deposits will be created by a Deposit service and will not
have a status until that time.

Finally, the userActionRequired has been replaced with a rejected status
option for Deposit.