Implementing [create|update]AndReadResource(...) convenience methods. (#32)
Adds two convenience methods to `PassClient`:
* `<T extends PassEntity> T createAndReadResource(T modelObj, Class<T> modelClass)`
* `<T extends PassEntity> T updateAndReadResource(T modelObj, Class<T> modelClass)`

Neither method is _guaranteed_ to return a resource that is the direct result of the creation or update (i.e. an interleaved write could occur between the `POST` or `PATCH`, and the subsequent `GET`).  However, `createAndReadResource(...)` is implemented in such a way (using `Prefer: return=representation`) that this guarantee holds.  Unfortunately, at this time, `updateAndReadResource(...)` does not, due to limitations of Fedora.