Adds a shaded artifact supporting model 2.3 (#55)
The shaded artifact `pass-client-shaded-v2_3` produces a jar whereby all client classes are prefixed by `org.dataconservancy.pass.v2_3`.  This allows simultaneous usage of the current model 3.x and model 2.3 in client code, supporting use cases that relate to converting/migrating 2.3 data to 3.x.  

Includes an integration test `ShadedClientIT` that tests and illustrates usage of multiple client versions for creating a 2.3 resource with a 2.3 client, then upgrading it with a 3.0 client.

Resolves #54