Add PUT option to Fedora updates (#56)
Adds optional parameter to `getPassClient(...)` to set `overwriteOnUpdate` to true. This allows you to, on instantiation of the `PassClient`, set it to always use `PUT` instead of of `PATCH` when you need to perform an update that removes any older versions of the model. By default `PATCH` will be used.

Some tidying up of Javadoc to fix formatting, and fix incorrect params on findByAttribute.